Dataframe containing socio economic data of the German electoral districts at the time of the 2017 federal elections (Bundestagswahl)

Data presented here has been altered in the sense that it has been rendered more machine-friendly (one header row only, no special characters, no blanks in header line, etc.). Beside that, the data itself has not been changed in any way.

Verwendung

data(socec)

Format

A data frame containing 316 rows and 51 columns

V01

state

V02

Wahlkreis-Nr.

V03

Wahlkreis-Name

V04

Gemeinden am 31.12.2015 (Anzahl)

V05

Flaeche am 31.12.2015 (km^2)

V06

Bevoelkerung am 31.12.2015 - Insgesamt (in 1000)

V07

Bevoelkerung am 31.12.2015 - Deutsche (in 1000)

V08

Bevoelkerung am 31.12.2015 - Auslaender in Prozent

V09

Bevoelkerungsdichte am 31.12.2015 (Einwohner je km²)

V10

Zu- (+) bzw. Abnahme (-) der Bevoelkerung 2015 - Geburtensaldo (je 1000 Einwohner)

V11

Zu- (+) bzw. Abnahme (-) der Bevoelkerung 2015 - Wanderungssaldo (je 1000 Einwohner)

V12

Alter von ... bis ... Jahren am 31.12.2015 - unter 18 in Prozent

V13

Alter von ... bis ... Jahren am 31.12.2015 - 18-24 in Prozent

V14

Alter von ... bis ... Jahren am 31.12.2015 - 25-34 in Prozent

V15

Alter von ... bis ... Jahren am 31.12.2015 - 35-59 in Prozent

V16

Alter von ... bis ... Jahren am 31.12.2015 - 60-74 in Prozent

V17

Alter von ... bis ... Jahren am 31.12.2015 - 75 und mehr in Prozent

V18

Zensus 2011, Bevoelkerung nach Migrationshintergrund am 09.05.2011 - ohne Migrationshintergrund in Prozent

V19

Zensus 2011, Bevoelkerung nach Migrationshintergrund am 09.05.2011 - mit Migrationshintergrund in Prozent

V20

Zensus 2011, Bevoelkerung nach Religionszugehoerigkeit am 09.05.2011 - Roemisch-katholische Kirche in Prozent

V21

"Zensus 2011, Bevoelkerung nach Religionszugehoerigkeit am 09.05.2011 - Evangelische Kirche in Prozent

V22

Zensus 2011, Bevoelkerung nach Religionszugehoerigkeit am 09.05.2011 - Sonstige, keine, ohne Angabe in Prozent

V23

Zensus 2011, Wohnungen in Wohngebaeuden am 09.05.2011 - Eigentuemerquote

V24

Bautaetigkeit und Wohnungswesen - Fertiggestellte Wohnungen 2014 (je 1000 Einwohner)

V25

Bautaetigkeit und Wohnungswesen - Bestand an Wohnungen am 31.12.2015 (je 1000 Einwohner)

V26

Verfuegbares Einkommen der privaten Haushalte 2014 (EUR je Einwohner)

V27

Bruttoinlandsprodukt 2014 (EUR je Einwohner)

V28

Kraftfahrzeugbestand am 01.01.2016 (je 1000 Einwohner)

V29

Absolventen/Abgaenger beruflicher Schulen 2015

V30

Absolventen/Abgaenger allgemeinbildender Schulen 2015 - insgesamt ohne Externe (je 1000 Einwohner)

V31

Absolventen/Abgaenger allgemeinbildender Schulen 2015 - ohne Hauptschulabschluss in Prozent

V32

Absolventen/Abgaenger allgemeinbildender Schulen 2015 - mit Hauptschulabschluss in Prozent

V33

Absolventen/Abgaenger allgemeinbildender Schulen 2015 - mit mittlerem Schulabschluss in Prozent

V34

Absolventen/Abgaenger allgemeinbildender Schulen 2015 - mit allgemeiner und Fachhochschulreife in Prozent

V35

Kindertagesbetreuung: Betreute Kinder am 01.03.2016 (je 1000 Einwohner)

V36

Unternehmensregister 2014 - Unternehmen insgesamt (je 1000 Einwohner)

V37

Unternehmensregister 2014 - Handwerksunternehmen (je 1000 Einwohner)

V38

Sozialversicherungspflichtig Beschaeftigte am 30.06.2016 - insgesamt (je 1000 Einwohner)

V39

Sozialversicherungspflichtig Beschaeftigte am 30.06.2016 - Land- und Forstwirtschaft, Fischerei in Prozent

V40

Sozialversicherungspflichtig Beschaeftigte am 30.06.2016 - Produzierendes Gewerbe in Prozent

V41

Sozialversicherungspflichtig Beschaeftigte am 30.06.2016 - Handel, Gastgewerbe, Verkehr in Prozent

V42

Sozialversicherungspflichtig Beschaeftigte am 30.06.2016 - oeffentliche und private Dienstleister in Prozent

V43

Sozialversicherungspflichtig Beschaeftigte am 30.06.2016 - uebrige Dienstleister und 'ohne Angabe' in Prozent

V44

Empfaenger(innen) von Leistungen nach SGB II am 31.12.2016 - insgesamt (je 1000 Einwohner)

V45

Empfaenger(innen) von Leistungen nach SGB II am 31.12.2016 - nicht erwerbsfaehige Hilfebeduerftige in Prozent

V46

Empfaenger(innen) von Leistungen nach SGB II am 31.12.2016 - Auslaender in Prozent

V47

Arbeitslosenquote Maerz 2017 - insgesamt

V48

Arbeitslosenquote Maerz 2017 - Maenner

V49

Arbeitslosenquote Maerz 2017 - Frauen

V50

Arbeitslosenquote Maerz 2017 - 15 bis unter 20 Jahre

V51

Arbeitslosenquote Maerz 2017 - 55 bis unter 65 Jahre

Quelle

This dataset is published by the Bundeswahlleiter 2017, (c) Der Bundeswahlleiter, Wiesbaden 2017, https://www.bundeswahlleiter.de/info/impressum.html, Data is made available for unrestricted use provided the source is credited.